change the directory of tennis plugin
